Add class for representing execution environments
The Environment class serves as a representation of arbitrarily
nested execution environments which implements the MutableMapping
protocol in a special way:

The __getitem__, __setitem__, __delitem__, __contains__, __iter__,
and __len__ methods are implemented in such a way that children
environments always call to the top-most parent environment.

As such, "immediate" versions these methods are also implemented
as regular methods prefixed with `i`, namely: iget, iset, idel,
ihas, iiter, and ilen respectively.