Make Symbol a Generic parameterized by an Amalgam
This changes Symbol to be a Generic that is parameterized by an Amalgam;
as such, the return type of the evaluate method for Symbol is also
modified. This also makes the type signature for SExpression be more
explicit especially for the func attribute that is to be evaluated to a
Function and then called.