Overhaul the Function Amalgam
This commit refactors the Function Amalgam in order to add a better
interface when used in tandem with the SExpression Amalgam. The
following changes would allow for an Environment to cascade and bind
to Functions properly when Amalgams are evaluated.

In order the achieve this result, the following changes have been
implemented for Function and the helper function create_fn:

- The bind protocol has been reintroduced.
- The evaluate method now only return self.
- The call method now accepts an environment.
- The closure_fn function now calls bind.